build: fix build on Solaris
The current binding.gyp file adds the same NAN headers directory to the
compiler's command line for Linux, Windows and OSX, but not for Solaris.

This results in node-protobuf not building on Solaris and platforms
derived from Solaris, like SmartOS.

This change fixes this problem by always adding NAN's headers directory
to the compiler's command line, regardless of the platform.